Refactor autocomplete to separate component
