+++ src/commons/navigationBar/subcomponents/SicpNavigationBar.test.tsx
@@ -1,3 +1,4 @@
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { render } from '@testing-library/react';
 import { MemoryRouter } from 'react-router';
 import { expect, test, vi } from 'vitest';
@@ -11,10 +12,15 @@ vi.mock('react-router', async () => ({
 }));
 
 test('Navbar renders correctly', () => {
+  const queryClient = new QueryClient({
+    defaultOptions: { queries: { retry: false } },
+  });
   const navbar = (
-    <MemoryRouter>
-      <SicpNavigationBar />
-    </MemoryRouter>
+    <QueryClientProvider client={queryClient}>
+      <MemoryRouter>
+        <SicpNavigationBar />
+      </MemoryRouter>
+    </QueryClientProvider>
   );
   const tree = render(navbar);
   expect(tree.asFragment()).toMatchSnapshot();
